refactor split up dictionary.ts
